Custom error response

+++ TourismSmartTransportation.Business/Implements/Admin/PartnerManagementService.cs
@@ -12,6 +12,7 @@ using TourismSmartTransportation.Data.Interfaces;
 using TourismSmartTransportation.Business.Extensions;
 using Azure.Storage.Blobs;
 using TourismSmartTransportation.Data.Models;
+using TourismSmartTransportation.Business.CommonModel;
 
 namespace TourismSmartTransportation.Business.Implements.Admin
 {
@@ -26,44 +27,47 @@ namespace TourismSmartTransportation.Business.Implements.Admin
         /// </summary>
         /// <param name="model"></param>
         /// <returns></returns>
-        public async Task<bool> AddPartner(AddPartnerModel model)
+        public async Task<Response> AddPartner(AddPartnerModel model)
         {
             bool isExist = await _unitOfWork.PartnerRepository.Query()
                 .AnyAsync(x => x.Username == model.Username);
             if (isExist)
             {
-                return false;
-            }
-            try
-            {
-                CreatePasswordHash(model.Password, out byte[] passwordHash, out byte[] passwordSalt);
-                var partner = new Partner()
+                return new()
                 {
-                    FirstName = model.FirstName,
-                    LastName = model.LastName,
-                    CompanyName = model.CompanyName,
-                    Address1 = model.Address1,
-                    Address2 = model.Address2,
-                    Phone = model.Phone,
-                    DateOfBirth = model.DateOfBirth != null ? model.DateOfBirth.Value : null,
-                    Gender = model.Gender,
-                    Email = model.Email,
-                    Password = passwordHash,
-                    Salt = passwordSalt,
-                    CreatedDate = DateTime.Now,
-                    ModifiedDate = DateTime.Now,
-                    Username = model.Username,
-                    PhotoUrl = UploadFile(model.UploadFile, Container.Partner).Result,
-                    Status = 1
+                    StatusCode = 400,
+                    Message = "Tài khoản đăng nhập đã tồn tại!"
                 };
-                await _unitOfWork.PartnerRepository.Add(partner);
-                await _unitOfWork.SaveChangesAsync();
             }
-            catch (Exception)
+
+            CreatePasswordHash(model.Password, out byte[] passwordHash, out byte[] passwordSalt);
+            var partner = new Partner()
             {
-                return false;
-            }
-            return true;
+                FirstName = model.FirstName,
+                LastName = model.LastName,
+                CompanyName = model.CompanyName,
+                Address1 = model.Address1,
+                Address2 = model.Address2,
+                Phone = model.Phone,
+                DateOfBirth = model.DateOfBirth != null ? model.DateOfBirth.Value : null,
+                Gender = model.Gender,
+                Email = model.Email,
+                Password = passwordHash,
+                Salt = passwordSalt,
+                CreatedDate = DateTime.Now,
+                ModifiedDate = DateTime.Now,
+                Username = model.Username,
+                PhotoUrl = UploadFile(model.UploadFile, Container.Partner).Result,
+                Status = 1
+            };
+            await _unitOfWork.PartnerRepository.Add(partner);
+            await _unitOfWork.SaveChangesAsync();
+
+            return new()
+            {
+                StatusCode = 201,
+                Message = "Tạo tài khoản thành công!"
+            };
         }
 
         /// <summary>
@@ -72,20 +76,28 @@ namespace TourismSmartTransportation.Business.Implements.Admin
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<bool> DeletePartner(Guid id)
+        public async Task<Response> DeletePartner(Guid id)
         {
-            try
+
+            var partner = await _unitOfWork.PartnerRepository.GetById(id);
+            if (partner == null)
             {
-                var Partner = await _unitOfWork.PartnerRepository.GetById(id);
-                Partner.Status = 0;
-                _unitOfWork.PartnerRepository.Update(Partner);
-                await _unitOfWork.SaveChangesAsync();
+                return new()
+                {
+                    StatusCode = 404,
+                    Message = "Không tìm thấy!"
+                };
             }
-            catch
+
+            partner.Status = 0;
+            _unitOfWork.PartnerRepository.Update(partner);
+            await _unitOfWork.SaveChangesAsync();
+
+            return new()
             {
-                return false;
-            }
-            return true;
+                StatusCode = 201,
+                Message = "Cập nhật trạng thái thành công!"
+            };
         }
 
         /// <summary>
@@ -137,38 +149,44 @@ namespace TourismSmartTransportation.Business.Implements.Admin
         /// <param name="id"></param>
         /// <param name="model"></param>
         /// <returns></returns>
-        public async Task<bool> UpdatePartner(Guid id, UpdatePartnerModel model)
+        public async Task<Response> UpdatePartner(Guid id, UpdatePartnerModel model)
         {
-            try
+            var partner = await _unitOfWork.PartnerRepository.GetById(id);
+            if (partner == null)
             {
-                var partner = await _unitOfWork.PartnerRepository.GetById(id);
-                if (model.Password != null)
+                return new()
                 {
-                    CreatePasswordHash(model.Password, out byte[] passwordHash, out byte[] passwordSalt);
-                    partner.Password = UpdateTypeOfNullAbleObject<byte[]>(partner.Password, passwordHash);
-                    partner.Salt = UpdateTypeOfNullAbleObject<byte[]>(partner.Salt, passwordSalt);
-                }
-                partner.PhotoUrl = await DeleteFile(model.DeleteFile, Container.Partner, partner.PhotoUrl);
-                partner.PhotoUrl += await UploadFile(model.UploadFile, Container.Partner);
-                partner.FirstName = UpdateTypeOfNullAbleObject<string>(partner.FirstName, model.FirstName);
-                partner.LastName = UpdateTypeOfNullAbleObject<string>(partner.LastName, model.LastName);
-                partner.Address1 = UpdateTypeOfNullAbleObject<string>(partner.Address1, model.Address1);
-                partner.Address2 = UpdateTypeOfNullAbleObject<string>(partner.Address2, model.Address2);
-                partner.Phone = UpdateTypeOfNullAbleObject<string>(partner.Phone, model.Phone);
-                partner.DateOfBirth = UpdateTypeOfNotNullAbleObject<DateTime>(partner.DateOfBirth, model.DateOfBirth);
-                partner.Email = UpdateTypeOfNullAbleObject<string>(partner.Email, model.Email);
-                partner.CompanyName = UpdateTypeOfNullAbleObject<string>(partner.CompanyName, model.CompanyName);
-                partner.Gender = UpdateTypeOfNotNullAbleObject<bool>(partner.Gender, model.Gender);
-                partner.Status = UpdateTypeOfNotNullAbleObject<int>(partner.Status, model.Status);
-                partner.ModifiedDate = DateTime.Now;
-                _unitOfWork.PartnerRepository.Update(partner);
-                await _unitOfWork.SaveChangesAsync();
+                    StatusCode = 404,
+                    Message = "Không tìm thấy!"
+                };
             }
-            catch (Exception)
+            if (model.Password != null)
             {
-                return false;
+                CreatePasswordHash(model.Password, out byte[] passwordHash, out byte[] passwordSalt);
+                partner.Password = UpdateTypeOfNullAbleObject<byte[]>(partner.Password, passwordHash);
+                partner.Salt = UpdateTypeOfNullAbleObject<byte[]>(partner.Salt, passwordSalt);
             }
-            return true;
+            partner.PhotoUrl = await DeleteFile(model.DeleteFile, Container.Partner, partner.PhotoUrl);
+            partner.PhotoUrl += await UploadFile(model.UploadFile, Container.Partner);
+            partner.FirstName = UpdateTypeOfNullAbleObject<string>(partner.FirstName, model.FirstName);
+            partner.LastName = UpdateTypeOfNullAbleObject<string>(partner.LastName, model.LastName);
+            partner.Address1 = UpdateTypeOfNullAbleObject<string>(partner.Address1, model.Address1);
+            partner.Address2 = UpdateTypeOfNullAbleObject<string>(partner.Address2, model.Address2);
+            partner.Phone = UpdateTypeOfNullAbleObject<string>(partner.Phone, model.Phone);
+            partner.DateOfBirth = UpdateTypeOfNotNullAbleObject<DateTime>(partner.DateOfBirth, model.DateOfBirth);
+            partner.Email = UpdateTypeOfNullAbleObject<string>(partner.Email, model.Email);
+            partner.CompanyName = UpdateTypeOfNullAbleObject<string>(partner.CompanyName, model.CompanyName);
+            partner.Gender = UpdateTypeOfNotNullAbleObject<bool>(partner.Gender, model.Gender);
+            partner.Status = UpdateTypeOfNotNullAbleObject<int>(partner.Status, model.Status);
+            partner.ModifiedDate = DateTime.Now;
+            _unitOfWork.PartnerRepository.Update(partner);
+            await _unitOfWork.SaveChangesAsync();
+
+            return new()
+            {
+                StatusCode = 201,
+                Message = "Cập nhật thành công!"
+            };
         }
     }
 }
